% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add lines}
\usage{
add_lines(data, extract_length = 6e+05, window_size = 60000)
}
\arguments{
\item{data}{dataset to divide into lines}

\item{extract_length}{total length of extract in ms}

\item{window_size}{length of line in ms}
}
\value{
data set with 'line' column added
}
\description{
Add lines
}
