% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_timing}
\alias{calculate_timing}
\title{Calculate conversation properties}
\usage{
calculate_timing(data)
}
\arguments{
\item{data}{talkr data frame}
}
\value{
data frame containing the UIDs and calculated columns turn_duration, transition_time
}
\description{
A dataframe is generated with conversation properties related to timing.
This data is made for quality control purposes only, and does not contain
sophisticated transition calculation methods. For this, we refer to the
python package `scikit-talk`.
}
