% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_columns}
\alias{check_columns}
\title{Check the presence of necessary columns in a dataset}
\usage{
check_columns(data, columns)
}
\arguments{
\item{data}{dataset to check}

\item{columns}{a vector of column names that must be present}
}
\value{
nothing, but throws an error if a column is missing
}
\description{
Check the presence of necessary columns in a dataset
}
