% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_quality.R
\name{check_quality}
\alias{check_quality}
\title{Check the quality of a specific source}
\usage{
check_quality(data, source)
}
\arguments{
\item{data}{dataset including the columns source, begin, and end}

\item{source}{the source that is verified}
}
\value{
0 if the quality is good, 1 if the quality is bad
}
\description{
Quality measure that is considered: the total duration of utterances in ms
must not be equal to the number of utterances.
}
