% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_time}
\alias{check_time}
\title{Verify that timing columns are numeric and likely indicate milliseconds.}
\usage{
check_time(column, name)
}
\arguments{
\item{column}{vector with timing information}

\item{name}{name of the column}
}
\value{
nothing, but throws an error if the column is not numeric and warns if the column may not indicate milliseconds
}
\description{
Verify that timing columns are numeric and likely indicate milliseconds.
}
