% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-stack.R
\name{coord_stack}
\alias{coord_stack}
\title{Coordinate system that stacks a wide plot to multiple layers}
\usage{
coord_stack(width, ...)
}
\arguments{
\item{width}{Width of the stacked layers in seconds}
}
\description{
Coordinate system that stacks a wide plot to multiple layers
}
