% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_turn_plot.R
\name{create_turn_plot_modified}
\alias{create_turn_plot_modified}
\title{create turn plot for testing purposes}
\usage{
create_turn_plot_modified()
}
\description{
create turn plot for testing purposes
}
