% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_turn.R
\name{geom_turn}
\alias{geom_turn}
\title{Title}
\usage{
geom_turn(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
