% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{get_ifadv}
\alias{get_ifadv}
\title{Get IFADV data}
\usage{
get_ifadv()
}
\value{
A tibble containing the data
}
\description{
IFA Dialog Video corpus data
Available in the public repository:
https://github.com/elpaco-escience/ifadv
}
