% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-stack.R, R/geom_token.R, R/geom_turn.R
\docType{data}
\name{CoordStack}
\alias{CoordStack}
\alias{GeomToken}
\alias{GeomTurn}
\title{CoordStack}
\format{
An object of class \code{CoordStack} (inherits from \code{CoordCartesian}, \code{Coord}, \code{ggproto}, \code{gg}) of length 15.

An object of class \code{GeomToken} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomTurn} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
CoordStack

GeomToken

GeomTurn
}
\description{
CoordStack

GeomToken

GeomTurn
}
\keyword{datasets}
