% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{inspect_corpus}
\alias{inspect_corpus}
\title{Inspect corpus}
\usage{
inspect_corpus(d, d.tokens, lang = NULL, saveplot = F, allsources = F)
}
\arguments{
\item{d}{dataset}

\item{d.tokens}{tokens dataset}

\item{lang}{language}

\item{saveplot}{should the plot be saved}

\item{allsources}{should all sources be shown}
}
\description{
A function that gives us a quick and rich impression of a language or
corpus.
}
