% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_language.R
\name{inspect_language}
\alias{inspect_language}
\title{Inspect language}
\usage{
inspect_language(
  data_conv,
  data_tokens,
  lang,
  saveplot = FALSE,
  allsources = FALSE
)
}
\arguments{
\item{data_conv}{conversation dataset}

\item{data_tokens}{tokens dataset}

\item{lang}{language}

\item{saveplot}{should the plot be saved}

\item{allsources}{all sources}
}
\description{
Inspect language
}
