% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_language.R
\name{inspect_tokens}
\alias{inspect_tokens}
\title{Inspect tokens}
\usage{
inspect_tokens(data, lang = NULL, saveplot = FALSE, allsources = FALSE)
}
\arguments{
\item{data}{dataset}

\item{lang}{language}

\item{saveplot}{should the plot be saved}

\item{allsources}{all sources}
}
\description{
Inspect tokens
}
