% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_density}
\alias{plot_density}
\title{Make a density plot of a specific column}
\usage{
plot_density(
  data,
  colname,
  title = "Density",
  xlab = "value",
  ylab = "density"
)
}
\arguments{
\item{data}{data frame containing the column}

\item{colname}{column name for which the density should be plotted}

\item{title}{plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
recorded plot
}
\description{
Make a density plot of a specific column
}
