% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality.R
\name{plot_quality}
\alias{plot_quality}
\title{Check source quality by plotting timing data}
\usage{
plot_quality(data, source = "all", saveplot = FALSE)
}
\arguments{
\item{data}{talkr data frame}

\item{source}{source to be checked (default is "all", no source is selected)}

\item{saveplot}{save plot to file (default is FALSE)}
}
\value{
list of recorded plots
}
\description{
Check source quality by plotting timing data
}
