% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Make a scatter plot of two columns}
\usage{
plot_scatter(
  data,
  colname_x,
  colname_y,
  title = "Scatter",
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{data}{data frame containing the columns}

\item{colname_x}{name of column plotted on x-axis}

\item{colname_y}{name of column plotted on y-axis}

\item{title}{plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
recorded plot
}
\description{
Make a scatter plot of two columns
}
