% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_turns_tokens}
\alias{plot_turns_tokens}
\title{Plot a stretch of conversation}
\usage{
plot_turns_tokens(data, begin = 0, duration = 60, maxrank = 10, source = 1)
}
\arguments{
\item{data}{a talkr dataset}

\item{begin}{start time of the plot in seconds (defaults to 0)}

\item{duration}{duration of the plot in seconds (defaults to 60)}

\item{maxrank}{maximum rank of tokens to plot (defaults to 10)}

\item{source}{name or number of the source to plot (defaults to 1, plotting the first source in the data)}
}
\value{
plot object
}
\description{
Utterances are plotted as grey boxes, with frequently occurring tokens overlaid as colored dots.
}
