% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_ordinals.R
\name{replace_ordinals}
\alias{replace_ordinals}
\title{Replace Ordinal Numeric in a Command}
\usage{
replace_ordinals(cmd, max_n = 100)
}
\arguments{
\item{cmd}{Command string to replace ordinal values}

\item{max_n}{Maxium number to go over.  For example
if \code{max_n = 100} then \code{"one hundred one"} will not be
replaced}
}
\value{
A character string same length as \code{cmd}
}
\description{
Replace Ordinal Numeric in a Command
}
\examples{
cmd = c("sort the twelfth column", "filter the third row ")
replace_ordinals(cmd)
}
