% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_stats}
\alias{report_stats}
\title{Report corpus-level and conversation-level statistics}
\usage{
report_stats(data)
}
\arguments{
\item{data}{talkr dataset}
}
\description{
Basic conversation statistics are reported to the console:
- Corpus-level statistics, reporting on the dataset as a whole;
- Conversation-level statistics, reporting per source.
}
\details{
The input for this function must be a `talkr` dataset, containing
the columns `source`, `participant`, `begin`, and `end`. Time stamps in the
columns `begin` and `end` must be in milliseconds.
To easily transform a dataset to a `talkr` dataset, consult `talkr::init()`.
}
