% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_language.R
\name{summarize_all_data}
\alias{summarize_all_data}
\title{Summarize all data}
\usage{
summarize_all_data(data, lang)
}
\arguments{
\item{data}{dataset}

\item{lang}{language}
}
\description{
Summarize all data
}
