% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_language.R
\name{summarize_language_data}
\alias{summarize_language_data}
\title{Summarize the data for a specific language}
\usage{
summarize_language_data(data, lang)
}
\arguments{
\item{data}{dataset}

\item{lang}{language}
}
\description{
Summarize the data for a specific language
}
