% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_talk_count.R
\name{talk_count}
\alias{talk_count}
\alias{talk_count_expr}
\title{Talk to Count a Data set}
\usage{
talk_count(.data, cmd, verbose = FALSE, ...)

talk_count_expr(data_colnames, cmd, allowed_words = "", ...)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}

\item{allowed_words}{words allowed to be in the command other than
the column names}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Count a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")
  cmds = c(
    "count by  mpg",
    "count up mpg, hp",
    "count up mpg, hp, mpg")
 data_colnames = df
 .data = df
 results = lapply(cmds, talk_count, .data = df)
df = df \%>\%
  rename(GEAR = gear)
data_colnames = df
cmd = "count by gear "
 gear = df \%>\%
 talk_count("count by gear ")
testthat::expect_true(all(colnames(gear) == c("GEAR", "n")))
}
