% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_dplyr_function.R
\name{talk_dplyr_function}
\alias{talk_dplyr_function}
\title{Determine the `dplyr` function from the command}
\usage{
talk_dplyr_function(cmd)
}
\arguments{
\item{cmd}{A command to pass, that will determine which
\code{dplyr} function to use}
}
\value{
A string
}
\description{
Determine the `dplyr` function from the command
}
\examples{
talk_dplyr_function( "Sort df by  mpg", "arrange the data by mpg")
}
