% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_filter.R
\name{talk_filter}
\alias{talk_filter}
\alias{talk_filter_expr}
\alias{talk_process_filter_cmd}
\alias{talk_colnames_class}
\title{Talk to Filter a Data set}
\usage{
talk_filter(.data, cmd, verbose = FALSE, ...)

talk_filter_expr(data_colnames, cmd, ...)

talk_process_filter_cmd(cmd)

talk_colnames_class(.data)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Filter a Data set
}
\examples{
library(tibble)
library(dplyr)
df = tibble::rownames_to_column(mtcars, var = "car")
df = df \%>\%
  rename(cylinders = cyl,
horsepower = hp,
American = am,
carburetor = carb)
cmds = c("subset rows, by horsepower > 100",
        "keep rows where horsepower less than 100",
        "drop rows where American equals 1",
        "drop rows where American is equal to 1",
        "drop rows where American is equal 1",
        "drop values in column 6 not equal to 4",
        "filter out MPG greater than 30",
        "filter the rows where American is not less than or equal 0",
        "return the data with values of cylinders not less than 4",
        "filter where cylinders are not missing",
        "filter if cylinders isn't missing",
        "filter so gear equal 6")
 data_colnames = colnames(df)
 exprs = lapply(cmds, talk_filter_expr, data_colnames = df)
 exprs = sapply(exprs, function(x) x$condition)
 results = lapply(cmds, talk_filter, .data = df)
 cyl = df \%>\%
 talk_filter("filter if cylinders is equal to 6")
testthat::expect_true(all(cyl$cylinders == 6))
 results = lapply(cmds, talk_filter, .data = df)
cmd = "filter where cylinders are not missing and gear greater than 6"
res = talk_filter_expr(cmd, data_colnames = df)
testthat::expect_true(res$condition == "!is.na(cylinders) & gear > 6")

cmd = "filter where cylinders equal to 4 or gear equals 6"
res = talk_filter_expr(cmd, data_colnames = df)
testthat::expect_true(res$condition == "cylinders == 4 | gear == 6")
talk_colnames_class(cyl)
}
