% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_get_colnames.R
\name{talk_get_colnames}
\alias{talk_get_colnames}
\alias{talk_cmd_to_colnames}
\alias{talk_check_colnames}
\title{Talk to Arrange a Data set}
\usage{
talk_get_colnames(data_colnames, cmd,
  stop_words = talkr::talk_stop_words, additional_stop_words = NULL,
  remove_duplicated = TRUE, allowed_words = c("descending",
  "ascending"), ...)

talk_cmd_to_colnames(data_colnames, cmd,
  stop_words = talkr::talk_stop_words, additional_stop_words = NULL)

talk_check_colnames(data_colnames)
}
\arguments{
\item{data_colnames}{column names of the data}

\item{cmd}{Command to perform on the data set.}

\item{stop_words}{Words to remove from the command}

\item{additional_stop_words}{additional stop words to
remove.  Helpful if you want to pass in these words
instead of with \code{stop_words}}

\item{remove_duplicated}{remove duplicated columns in the
command}

\item{allowed_words}{words allowed to be in the command other than
the column names}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Arrange a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")
  cmds = c(
    "Sort by  mpg",
    "Sort by  column    mpg  ",
    "arrange by column 5",
    "arrange by gear",
    "arrange by columns 4 and 5",
    "arrange by columns 4 and 5, mpg decreasing",
    # duplciate
    "arrange by columns 2 and 5, mpg decreasing",
    "arrange by columns 4, 5, and 6",
    "sort by mpg descending",
    "sort by mpg ascending",
    "sort by mpg ascending",
    "sort by mpg low to high hat")
 data_colnames = df
 cmd = "arrange by columns 4 and 5, mpg decreasing"
 results = lapply(cmds, talk_get_colnames, data_colnames = df)
 results = lapply(cmds, talk_get_colnames, data_colnames = df,
 allowed_words = "")
 results = lapply(cmds, talk_get_colnames, data_colnames = df,
 allowed_words = NULL)

}
