% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_qplot.R
\name{talk_ggplot}
\alias{talk_ggplot}
\alias{talk_ggplot_expr}
\alias{talk_process_ggplot_cmd}
\title{Talk to ggplot a Data set}
\usage{
talk_ggplot(.data, cmd, verbose = FALSE, ...)

talk_ggplot_expr(data_colnames, cmd, allowed_words = talk_ggplot_words(),
  ...)

talk_process_ggplot_cmd(cmd)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to ggplot a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")

  cmds = c(
    "ggplot by  mpg",
    "ggplot by  column    mpg  ",
    "ggplot by column 5",
    "ggplot a histogram of mpg, coloured by gear",
    "ggplot columns 4 and 5",
    # duplciate
    "ggplot by mpg coloured by American")
 data_colnames = df
 .data = df
 allowed_words = talkr:::talk_ggplot_words()
 cmd = c("ggplot a histogram of mpg, coloured by gear",
 "ggplot a histogram of mpg, facetted by gear")
 results = lapply(cmds, talk_ggplot_expr, data_colnames = df)
 results = lapply(cmd, talk_ggplot, .data = df)
 cmd =  "ggplot by columns 2 and 5, mpg decreasing"
 testthat::expect_warning(talk_ggplot(.data, cmd),
 "orderings")
df = df \%>\%
  rename(GEAR = gear)
 gear = df \%>\%
 talk_ggplot("arrange by gear and column 3")
testthat::expect_true(all(ggplot_vars(gear) == c("GEAR", "cyl")))
}
