% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_group_by.R
\name{talk_group_by}
\alias{talk_group_by}
\alias{talk_group_by_expr}
\alias{talk_process_group_by_cmd}
\title{Talk to Group a Data set}
\usage{
talk_group_by(.data, cmd, verbose = FALSE, ...)

talk_group_by_expr(data_colnames, cmd, ...)

talk_process_group_by_cmd(cmd)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Group a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")
  cmds = c(
    "group by  mpg",
    "Group by  column    mpg  ",
    "group by column 5",
    "arrange by gear",
    "group by columns 4 and 5",
    "group by columns 4 and 5, mpg decreasing",
    # duplciate
    "group by columns 2 and 5, mpg decreasing",
    "group by columns 4, 5, and 6",
    "group by mpg descending",
    "group by mpg ascending",
    "group by mpg ascending",
    "group by mpg low to high")
 data_colnames = df
 .data = df
 results = lapply(cmds, talk_group_by, .data = df)
df = df \%>\%
  rename(GEAR = gear)
 gear = df \%>\%
 talk_group_by("arrange by gear and column 3")
testthat::expect_true(all(group_vars(gear) == c("GEAR", "cyl")))
}
