% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_replace_all.R
\name{talk_replace}
\alias{talk_replace}
\alias{talk_replace_expr}
\alias{talk_process_replace_cmd}
\title{Talk to Replace all types a Data set}
\usage{
talk_replace(.data, cmd, verbose = FALSE, ...)

talk_replace_expr(
  data_colnames,
  cmd,
  allowed_words = c("factor", "logical", "numeric", "character", "double"),
  ...
)

talk_process_replace_cmd(cmd)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}

\item{allowed_words}{words allowed to be in the command other than
the column names}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Replace all types a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")
  cmds = c(
    "replace gear mpg as factors",
    "replace car as factor",
    "replace gear as character")
 data_colnames = df
 .data = df
 cmd = cmds[1]
 results = lapply(cmds, talk_replace, .data = df)
 testthat::expect_warning(talk_replace(.data, cmd),
 "allowed")
df = df \%>\%
  rename(GEAR = gear)
 gear = df \%>\%
 talk_replace("replace by gear and column 3 as factors")
 d_classes = sapply(gear[, c("cyl", "GEAR")], class)
testthat::expect_true(all(d_classes == "factor"))
}
