% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_select.R
\name{talk_select}
\alias{talk_select}
\alias{talk_select_expr}
\alias{talk_process_select_cmd}
\title{Talk to Filter a Data set}
\usage{
talk_select(.data, cmd, verbose = FALSE, ...)

talk_select_expr(data_colnames, cmd, ...)

talk_process_select_cmd(cmd)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Filter a Data set
}
\examples{
library(tibble)
library(dplyr)
df = tibble::rownames_to_column(mtcars, var = "car")
df = df \%>\%
  rename(cylinders = cyl,
horsepower = hp,
American = am,
carburetor = carb)
.data = df
cmds = c("subset columns 5 and 6",
        "select column horsepower",
        "drop columns 1 and 3 american",
        "select MPG column four 5")
cmd = cmds
 data_colnames = colnames(df)
 exprs = sapply(cmds, talk_select_expr, data_colnames = df)
 exprs = sapply(exprs, function(x) x$condition)
 results = lapply(cmds, talk_select, .data = df)
 cyl = df \%>\%
 talk_select("select if cylinders American")
testthat::expect_true(all(colnames(cyl) == c("cylinders", "American")))
}
