% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_shiny_example.R
\name{talk_shiny_example}
\alias{talk_shiny_example}
\title{Talk Shiny Examples}
\usage{
talk_shiny_example(example = c("sort_better", "simple_sort"))
}
\arguments{
\item{example}{Name of example to run.}
}
\value{
The output from \code{\link{runApp}}
}
\description{
Talk Shiny Examples
}
\examples{
if (interactive()) {
talk_shiny_example()
}
}
