% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_talk_stop_words.R
\docType{data}
\name{talk_stop_words}
\alias{talk_stop_words}
\title{A character vector containing the common words to remove from a command for
\code{talkr}}
\format{An object of class \code{character} of length 20.}
\usage{
talk_stop_words
}
\description{
A character vector containing the common words to remove from a command for
\code{talkr}
}
\keyword{datasets}
