% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_turnPlot.R
\name{theme_turnPlot}
\alias{theme_turnPlot}
\title{T  heme for the turn plot}
\usage{
theme_turnPlot(base_size = 11, base_family = "serif", ticks = TRUE)
}
\arguments{
\item{base_size}{int}

\item{base_family}{chr}

\item{ticks}{bool}
}
\value{
ggplot2 custom theme for turn plots
}
\description{
T  heme for the turn plot
}
