% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add information for line-by-line visualization}
\usage{
add_lines(data, line_duration = 60000)
}
\arguments{
\item{data}{dataset to divide into lines}

\item{line_duration}{length of line in ms}
}
\value{
data set with added columns `line`, `begin_line`, `end_line`
}
\description{
This function adds columns to the dataset that adds a line ID, and a `begin_line` and
`end_line` column for timestamps relative to the beginning of the line, so data can
be visualized line-by-line.
}
