% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-stack.R
\name{coord_stack}
\alias{coord_stack}
\title{Coordinate system that stacks a wide plot to multiple layers}
\usage{
coord_stack(width, expand = TRUE, clip = "on")
}
\arguments{
\item{width}{the length of a single layer}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. In most cases, the default of \code{"on"} should not be changed,
as setting \code{clip = "off"} can cause unexpected results. It allows
drawing of data points anywhere on the plot, including in the plot margins. If
limits are set via \code{xlim} and \code{ylim} and some data points fall outside those
limits, then those data points may show up in places such as the axes, the
legend, the plot title, or the plot margins.}
}
\description{
Coordinate system that stacks a wide plot to multiple layers
}
