% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{get_ifadv}
\alias{get_ifadv}
\title{Get IFADV data}
\usage{
get_ifadv(destfile = "data/ifadv.rda", clean_after = TRUE)
}
\arguments{
\item{destfile}{(default = "data/ifadv.rda") Destination file}

\item{clean_after}{(default = TRUE) Cleans the downloaded file after using it}
}
\value{
Nothing, but it loads the data into the global environment
}
\description{
IFA Dialog Video corpus data
Available in the public repository:
https://github.com/elpaco-escience/ifadv
}
