% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{get_ifadv}
\alias{get_ifadv}
\title{Get IFADV data}
\usage{
get_ifadv(
  source = "https://raw.githubusercontent.com/elpaco-escience/ifadv/csv/data/ifadv.csv"
)
}
\arguments{
\item{source}{(default = "https://raw.githubusercontent.com/elpaco-escience/ifadv/csv/data/ifadv.csv")}
}
\value{
A data frame containing the IFADV dataset
}
\description{
IFA Dialog Video corpus data
Available in the public repository:
https://github.com/elpaco-escience/ifadv
}
\details{
This function requires an internet connection.
}
