% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_token.R, R/geom_turn.R
\docType{data}
\name{GeomToken}
\alias{GeomToken}
\alias{GeomTurn}
\title{GeomToken}
\format{
An object of class \code{GeomToken} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomTurn} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomToken

GeomTurn
}
\description{
GeomToken

GeomTurn
}
\keyword{datasets}
