% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a `talkr` dataset}
\usage{
init(
  data,
  begin = "begin",
  end = "end",
  participant = "participant",
  utterance = "utterance"
)
}
\arguments{
\item{data}{A dataframe object}

\item{begin}{The column name with the begin time of the utterance}

\item{end}{The column name with the end time of the utterance}

\item{participant}{The participant who produced the utterance}

\item{utterance}{The utterance itself}
}
\value{
A dataframe object with columns needed for the talkr workflow
}
\description{
From a dataframe object, generate a talkr dataset.
This dataset contains columns that are used throughout the talkr
infrastructure to visualize conversations and language corpora.
Initializing a talkr dataset is the first step in the talkr workflow.
}
