% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_language.R
\name{inspect_language}
\alias{inspect_language}
\title{Inspect language}
\usage{
inspect_language(
  data_conv,
  data_tokens = NULL,
  lang = NULL,
  returnplots = FALSE,
  saveplot = FALSE,
  allsources = FALSE
)
}
\arguments{
\item{data_conv}{conversation dataset}

\item{data_tokens}{tokens dataset}

\item{lang}{language}

\item{returnplots}{boolean indicating whether plots should be made}

\item{saveplot}{should the plot be saved}

\item{allsources}{all sources}
}
\description{
Inspect language
}
