% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_turns_tokens.R
\name{plot_turns_tokens}
\alias{plot_turns_tokens}
\title{Plot conversation}
\usage{
plot_turns_tokens(data, sourcecount = 1)
}
\arguments{
\item{data}{a talkr dataset}

\item{source}{a sourcename to extract; default is NULL, then the first source will be picked}
}
\value{
plot object
}
\description{
Plot conversation
}
