% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{report_summaries}
\alias{report_summaries}
\title{Report summaries on data}
\usage{
report_summaries(data, lang, allsources)
}
\arguments{
\item{data}{dataset}

\item{lang}{language}

\item{allsources}{print all sources}
}
\description{
Report summaries on data
}
