% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarize_source_data}
\alias{summarize_source_data}
\title{Summarize source data}
\usage{
summarize_source_data(data, lang)
}
\arguments{
\item{data}{dataset}

\item{lang}{language}
}
\description{
Summarize source data
}
