% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talk_arrange.R
\name{talk_arrange}
\alias{talk_arrange}
\alias{talk_arrange_expr}
\alias{talk_get_colnames}
\alias{talk_process_arrange_cmd}
\title{Talk to Arrange a Data set}
\usage{
talk_arrange(.data, cmd, verbose = FALSE, ...)

talk_arrange_expr(data_colnames, cmd, ...)

talk_get_colnames(data_colnames, cmd, stop_words = c("sort", "arrange",
  "order", "group", "filter", "mutate", "subset", "select", "count",
  "transmute", "summarize", "summarise", "missing", "infinite", "finite",
  "nan", "not a number"))

talk_process_arrange_cmd(cmd)
}
\arguments{
\item{.data}{The data set/\code{data.frame} to
perform the operation.}

\item{cmd}{Command to perform on the data set.}

\item{verbose}{print diagnostic output}

\item{...}{additional arguments to pass to \code{\link{talk_get_colnames}}}

\item{data_colnames}{column names of the data}

\item{stop_words}{Words to remove from the command}
}
\value{
A \code{data.frame} or list of them if \code{length(cmds) > 1}.
}
\description{
Talk to Arrange a Data set
}
\examples{
library(tibble)
library(dplyr)
df = mtcars \%>\%
  rownames_to_column(var = "car")
  cmds = c(
    "Sort by  mpg",
    "Sort by  column    mpg  ",
    "arrange by column 5",
    "arrange by gear",
    "arrange by columns 4 and 5",
    "arrange by columns 4 and 5, mpg decreasing",
    # duplciate
    "arrange by columns 2 and 5, mpg decreasing",
    "arrange by columns 4, 5, and 6",
    "sort by mpg descending",
    "sort by mpg ascending",
    "sort by mpg ascending",
    "sort by mpg low to high")
 results = lapply(cmds, talk_arrange, .data = df)
df = df \%>\%
  rename(GEAR = gear)
 gear = df \%>\%
 talk_arrange("arrange by gear")
testthat::expect_true(!is.unsorted(gear$GEAR))
}
